Authentication, route guards, authorization

+++ controllers/delete.js
@@ -3,8 +3,17 @@ const { deleteCar } = require("../services/cars");
 module.exports = {
   async delCar(req, res) {
     const id = req.params.id;
-    console.log(id);
-    await deleteCar(id);
-    res.redirect("/");
+    let isOwner = false;
+
+    if (req.session.user && req.session.user.id == car.owner) {
+      isOwner = true;
+    }
+    
+    if (isOwner) {
+      await deleteCar(id);
+      res.redirect("/");
+    } else {
+      res.redirect(`/details/${id}`);
+    }
   },
 };
